+++ DAGs/dag-files/gbif-spark-example.py
@@ -1,22 +1,25 @@
-from datetime import datetime, timedelta, timezone
-from operators import SparkKubernetesOperator
-from hooks import SparkKubernetesSensor
-from airflow.models import Variable
+from operators.stackable_spark_operator import SparkKubernetesOperator
+from sensors.stackable_spark_sensor import SparkKubernetesSensor
+from airflow.models.param import Param
 from airflow import DAG
+from datetime import timedelta
+import pendulum
 
 with DAG(
     dag_id='sparkapp_dag',
     schedule_interval='0 0 * * *',
-    start_date=datetime(year=2022, month=1, day=1, tzinfo=timezone.utc),
+    start_date=pendulum.datetime(2022, 5, 1, tz="Europe/Copenhagen"),
     catchup=False,
-    dagrun_timeout=timedelta(day=7),
+    dagrun_timeout=timedelta(minutes=60),
     tags=['spark'],
-    params={"namespace_to_run": "spark-demo"},
+    params= {
+        "namespace_to_run": Param("stack-demo")
+        },
 ) as dag:
 
     t1 = SparkKubernetesOperator(
         task_id='spark_map_submit',
-        namespace = Variable.get("namespace_to_run"),
+        namespace = "{{ params.namespace_to_run }}",
         application_file="gbif-map-proces.yaml",
         do_xcom_push=True,
         dag=dag,
@@ -24,7 +27,7 @@ with DAG(
 
     t2 = SparkKubernetesSensor(
         task_id='spark_map_monitor',
-        namespace = Variable.get("namespace_to_run"),
+        namespace = "{{ params.namespace_to_run }}",
         application_name="{{ task_instance.xcom_pull(task_ids='spark_map_submit')['metadata']['name'] }}",
         poke_interval=5,
         dag=dag,
